Refactored into separate hook for App.js
